put `S3MediaStorage` and `LocalMediaStorage` in separate files
