+++ piccolo_admin/media/__init__.py
@@ -1,0 +1,4 @@
+"""
+Resources:
+https://cheatsheetseries.owasp.org/cheatsheets/File_Upload_Cheat_Sheet.html
+"""
